pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relic"
version = "0.1.0"
description = "Relic — Local LLM-powered pentesting automation within virtualized environments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Relic Contributors"}
]
keywords = ["pentesting", "llm", "automation", "security", "virtualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "textual>=0.47.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "paramiko>=3.4.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
relic = "relic.cli:main"
relic-web = "relic.web.api:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
